set up celery beat
