+++ meerkat_abacus/test/db_test.py
@@ -6,7 +6,7 @@ from sqlalchemy.dialects.postgresql import JSONB
 from dateutil.parser import parse
 from datetime import datetime
 
-from meerkat_abacus import manage
+from meerkat_abacus import data_management as manage
 from meerkat_abacus import model
 from meerkat_abacus import config
 
@@ -43,11 +43,11 @@ class DbTest(unittest.TestCase):
                 assert r.parent_location == 2
             if r.id == 7:
                 assert r.deviceid == "1,6"
-
+        form_directory = config.data_directory + "forms/"
         manage.fake_data(config.country_config,
-                         config.form_directory, engine, N=500)
+                         form_directory, engine, N=500)
         manage.import_data(config.country_config,
-                           config.form_directory,
+                           form_directory,
                            engine)
         results = session.query(manage.form_tables["case"])
         assert len(results.all()) == 500
